ch32/tinyUSB: Fix prototype of dcd_int_handler

+++ libraries/tinyUsb/ch32v3x/dcd_usbhs.cpp
@@ -445,7 +445,7 @@ void dcd_int_handler(uint8_t rhport) {
 }
 
 
-extern "C" void dcd_int_handler(int port);
+extern "C" void dcd_int_handler(uint8_t port);
 void USBHS_IRQHandler()
 {
     dcd_int_handler(0);
